move default methods to struct impl
