+++ src/eth/storage/permanent_storage.rs
@@ -1,4 +1,3 @@
-use anyhow::anyhow;
 use async_trait::async_trait;
 
 use super::EthStorageError;
@@ -24,8 +23,6 @@ pub trait PermanentStorage: Send + Sync {
     // Retrieves the last mined block number.
     async fn read_current_block_number(&self) -> anyhow::Result<BlockNumber>;
 
-    /// Atomically increments the block number, returning the new value.
-    async fn increment_block_number(&self) -> anyhow::Result<BlockNumber>;
     /// Checks if the transaction execution conflicts with the current storage state.
     async fn check_conflicts(&self, execution: &Execution) -> anyhow::Result<Option<ExecutionConflicts>>;
 
@@ -59,51 +56,4 @@ pub trait PermanentStorage: Send + Sync {
     ///
     /// TODO: maybe can use save_block from a default method.
     async fn enable_genesis(&self, genesis: Block) -> anyhow::Result<()>;
-
-    // -------------------------------------------------------------------------
-    // Default operations
-    // -------------------------------------------------------------------------
-
-    /// Retrieves an account from the storage. Returns default value when not found.
-    async fn read_account(&self, address: &Address, point_in_time: &StoragePointInTime) -> anyhow::Result<Account> {
-        match self.maybe_read_account(address, point_in_time).await? {
-            Some(account) => Ok(account),
-            None => Ok(Account {
-                address: address.clone(),
-                ..Account::default()
-            }),
-        }
-    }
-
-    /// Retrieves an slot from the storage. Returns default value when not found.
-    async fn read_slot(&self, address: &Address, slot_index: &SlotIndex, point_in_time: &StoragePointInTime) -> anyhow::Result<Slot> {
-        match self.maybe_read_slot(address, slot_index, point_in_time).await? {
-            Some(slot) => Ok(slot),
-            None => Ok(Slot {
-                index: slot_index.clone(),
-                ..Default::default()
-            }),
-        }
-    }
-
-    /// Translates a block selection to a specific storage point-in-time indicator.
-    async fn translate_to_point_in_time(&self, block_selection: &BlockSelection) -> anyhow::Result<StoragePointInTime> {
-        match block_selection {
-            BlockSelection::Latest => Ok(StoragePointInTime::Present),
-            BlockSelection::Number(number) => {
-                let current_block = self.read_current_block_number().await?;
-                if number <= &current_block {
-                    Ok(StoragePointInTime::Past(*number))
-                } else {
-                    Ok(StoragePointInTime::Past(current_block))
-                }
-            }
-            BlockSelection::Earliest | BlockSelection::Hash(_) => match self.read_block(block_selection).await? {
-                Some(block) => Ok(StoragePointInTime::Past(block.header.number)),
-                None => Err(anyhow!(
-                    "failed to select block because it is greater than current block number or block hash is invalid."
-                )),
-            },
-        }
-    }
 }
