+++ src/eth/storage/temporary_storage.rs
@@ -1,7 +1,9 @@
 use async_trait::async_trait;
 
+use super::EthStorageError;
 use crate::eth::primitives::Account;
 use crate::eth::primitives::Address;
+use crate::eth::primitives::Block;
 use crate::eth::primitives::BlockNumber;
 use crate::eth::primitives::Execution;
 use crate::eth::primitives::ExecutionConflicts;
@@ -12,7 +14,7 @@ use crate::eth::primitives::StoragePointInTime;
 /// Temporary storage (in-between blocks) operations
 // TODO: add Metrified method
 #[async_trait]
-pub trait OverlayStorage: Send + Sync {
+pub trait TemporaryStorage: Send + Sync {
     // Retrieves the last mined block number.
     async fn read_current_block_number(&self) -> anyhow::Result<BlockNumber>;
 
@@ -29,37 +31,11 @@ pub trait OverlayStorage: Send + Sync {
     async fn maybe_read_slot(&self, address: &Address, slot_index: &SlotIndex, point_in_time: &StoragePointInTime) -> anyhow::Result<Option<Slot>>;
 
     /// Commits changes to permanent storage
-    async fn commit(&self, block_number: BlockNumber, execution: Execution) -> anyhow::Result<()>;
+    async fn commit(&self, block: Block) -> anyhow::Result<(), EthStorageError>;
 
     /// Temporarily stores account changes during block production
     async fn save_account_changes(&self, block_number: BlockNumber, execution: Execution) -> anyhow::Result<()>;
 
     /// Resets all state to a specific block number.
     async fn reset(&self, number: BlockNumber) -> anyhow::Result<()>;
-
-    // -------------------------------------------------------------------------
-    // Default operations
-    // -------------------------------------------------------------------------
-
-    /// Retrieves an account from the storage. Returns default value when not found.
-    async fn read_account(&self, address: &Address, point_in_time: &StoragePointInTime) -> anyhow::Result<Account> {
-        match self.maybe_read_account(address, point_in_time).await? {
-            Some(account) => Ok(account),
-            None => Ok(Account {
-                address: address.clone(),
-                ..Account::default()
-            }),
-        }
-    }
-
-    /// Retrieves an slot from the storage. Returns default value when not found.
-    async fn read_slot(&self, address: &Address, slot_index: &SlotIndex, point_in_time: &StoragePointInTime) -> anyhow::Result<Slot> {
-        match self.maybe_read_slot(address, slot_index, point_in_time).await? {
-            Some(slot) => Ok(slot),
-            None => Ok(Slot {
-                index: slot_index.clone(),
-                ..Default::default()
-            }),
-        }
-    }
 }
